// web/src/generated/api/models/memberResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipRole } from './membershipRole';

export interface MemberResponse {
  id: string;
  userId: string;
  fullName: string;
  email: string;
  role: MembershipRole;
  createdAt: string;
}